requirements.txt
sqlalchemy
boto3
botocore
psycopg2-binary
python-dotenv